+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - split help document
 - add some subcommands on v2
 - add `insdeltasec` subcommand
+- add `groupsum` subcommand
 
 ## version 0.2.0 (2018-02-25)
 - add -v1, -v2 options